changed self-filtering in rotunit_filter_chain
since the robot_self_filter is not supported in hydo, self filtering is now based on the laser_filter/LaserScanFootprintFilter which performs a radial point removal around the base_link. in terms of accuracy this is definetly a step back, so we see to get the robot_self_filter working agian.

furthermore the scan topics were changed to be scan360 and scan360_filtered to be better identifiable when rotunit and regular 2d lasers are used side by side